feat: replace bmp280 sim with superset bmp3xx sim
