+++ config/config.exs
@@ -12,8 +12,7 @@ config :circuits_sim,
     {CircuitsSim.Device.SHT4X, bus_name: "i2c-1", address: 0x44},
     {CircuitsSim.Device.VEML7700, bus_name: "i2c-1", address: 0x48},
     {CircuitsSim.Device.SGP30, bus_name: "i2c-1", address: 0x58},
-    {CircuitsSim.Device.BMP280, bus_name: "i2c-1", address: 0x76},
-    {CircuitsSim.Device.BMP280, bus_name: "i2c-1", address: 0x77},
+    {CircuitsSim.Device.BMP3XX, bus_name: "i2c-1", address: 0x77},
     {CircuitsSim.Device.TM1620, bus_name: "spidev0.0", render: :binary_clock},
     {CircuitsSim.Device.GPIOLED, pin_spec: 10},
     {CircuitsSim.Device.GPIOButton, pin_spec: 11}
